# backend/openapi.yaml
openapi: 3.1.0
info:
  title: ProtoCMS API
  description: |
    A prototype in-memory headless CMS built with Go 1.26 stdlib.
    Supports dynamic content types and flexible content items persisted to JSON on disk.
  version: 1.0.0
  contact: {}
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/datasets/{name}/load:
    parameters:
      - $ref: '#/components/parameters/datasetName'

    post:
      summary: Load a dataset into memory
      operationId: loadDataset
      tags: [Datasets]
      description: Requires a bearer token with the `admin` role.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dataset loaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/uploads:
    post:
      summary: Upload an image to the caller's dataset
      operationId: uploadFile
      tags: [Uploads]
      description: |
        Stores an uploaded image in the uploads folder of the dataset bound to
        the credential (`data/<dataset>/uploads/`). Only migrated (v2) datasets
        support uploads. Accepts `multipart/form-data` with a single `file`
        field; max size 8 MiB. Allowed types: png, jpg/jpeg, gif, webp, svg.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: File stored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResult'
        '400':
          description: Missing/invalid file, unsupported type, or too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Dataset is in legacy format; migrate it to enable uploads
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/uploads/{dataset}/{name}:
    parameters:
      - $ref: '#/components/parameters/datasetName'
      - name: name
        in: path
        required: true
        description: The stored file name returned by the upload response.
        schema:
          type: string
          example: a5e58b4cd933d3548b29c9dab33daf3e.png

    get:
      summary: Serve an uploaded image
      operationId: serveUpload
      tags: [Uploads]
      description: |
        Public (no authentication) so plain `<img>` tags can load images. The
        dataset is taken from the path. Responds with the image bytes and a
        long-lived immutable cache header.
      responses:
        '200':
          description: The image file
          content:
            image/png: {}
            image/jpeg: {}
            image/gif: {}
            image/webp: {}
            image/svg+xml: {}
        '400':
          description: Invalid dataset or file name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Dataset is in legacy format; migrate it to enable uploads
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        A bearer token that is either a static API key (configured via
        `PROTOCMS_API_KEYS`) or an HS256 JWT obtained from `POST /api/login`.
        The token's role (`admin` or `editor`) governs which writes are allowed.
  responses:
    Unauthorized:
      description: Missing, malformed, or invalid bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated, but the token's role lacks permission
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
...
    datasetName:
      name: name
      in: path
      required: true
      description: Name of the dataset
      schema:
        type: string
        example: default
  schemas:
...
    DatasetStats:
      type: object
      properties:
        dataset:
          type: string
          description: Name of the active dataset
          example: default
        content_types:
          type: integer
          description: Number of registered content type schemas
          example: 1
        total_items:
          type: integer
          description: Total number of content items across all types
          example: 2
        items_per_type:
          type: object
          description: Item count broken down by content type name
          additionalProperties:
            type: integer
          example:
            post: 2
    Metadata:
      type: object
      description: Editable and system metadata for a dataset
      properties:
        name:
          type: string
          example: default
        author:
          type: string
          example: jane
        description:
          type: string
          example: Primary content dataset
        created_at:
          type: string
          format: date-time
          example: "2026-06-13T10:00:00Z"
        modified_at:
          type: string
          format: date-time
          example: "2026-06-19T12:30:00Z"
        data_format_version:
          type: integer
          example: 2
        schema_version:
          type: integer
          example: 1
        tags:
          type: array
          items:
            type: string
          example: [prod, blog]
...
    MetricsReport:
      type: object
      description: Aggregated query metrics for a dataset
      properties:
        total_queries:
          type: integer
          description: Total number of queries recorded
          example: 42
        by_op:
          type: object
          description: |
            Map of operation name (`list`, `get`, `create`, `update`,
            `delete`, `filter`) → map of content-type name → timing stats.
          additionalProperties:
            type: object
            additionalProperties:
              type: object
              properties:
                count:
                  type: integer
                  example: 10
                avg_ms:
                  type: number
                  example: 0.42
                total_ms:
                  type: number
                  example: 4.2
          example:
            list:
              post:
                count: 10
                avg_ms: 0.42
                total_ms: 4.2
    DatasetInfo:
      type: object
      description: Full runtime info for a single dataset
      properties:
        name:
          type: string
          example: default
        data_format_version:
          type: integer
          example: 2
        meta:
          $ref: '#/components/schemas/Metadata'
        stats:
          $ref: '#/components/schemas/DatasetStats'
        approx_bytes:
          type: integer
          description: Estimated in-memory size of the dataset in bytes
          example: 4096
        metrics:
          $ref: '#/components/schemas/MetricsReport'
    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
          example: not found
    UploadResult:
      type: object
      properties:
        url:
          type: string
          description: Path to serve the file, including the dataset segment.
          example: /api/uploads/shop/a5e58b4cd933d3548b29c9dab33daf3e.png
        name:
          type: string
          description: The stored (randomized) file name.
          example: a5e58b4cd933d3548b29c9dab33daf3e.png
        original:
          type: string
          description: The original client-supplied filename.
          example: logo.png